Reformat Raw8 to be a dataclass
Breaking Change!

Rename functions to be in snake_case instead of camelCase and make use
of the dataclass property decorator to rename getter functions to appear
more like object attributes. This simplifies access, but will break any
existing scripts that relied on the old access methods.

For intance, instead of accessing the wavelength through a function
getWavelength(), the wavelength can be read as a property of the object.

A statement that gathered the wavelength values from the class object
used to be written like the following:

x_data = raw8_object.getWavelength()

With this commit, it is now accessed according to the following syntax:

x_data = raw8_object.wavelength