Add attribute access support for header fields
Sanitize any byte strings in the header with an assumed "latin1"
decoding scheme.

Write dataclass properties to unpack the remaining fields in the header
so that they can be accessed directly from the object.

For instance, users can use the notation raw8_object.comment instead of
the original unwieldy raw8_object.header["comment"]. The new syntax
allows code editors to take advantage of autocompletion tools to access
header attributes instead of forcing the user to remember the available
fields inside the header dictionary.

Note: "latin1" decoding is assumed because the userfriendlyname field
appears to have special characters when decoded with utf-8. I suspect
the userfriendlyname field is unpacked to the wrong size. If unpacked
correctly, the proper decoding mechanism might be utf-8.